Use memory optimizer for both register and general-purpose memory.
